v2: merge .core and .problem
Move `v2.problem.*` to `v2.core`. Keep all pydantic models in the same module. Avoid unnecessary import complexity.
