+++ petab/v2/lint.py
@@ -15,8 +15,7 @@ import pandas as pd
 import sympy as sp
 
 from ..v2.C import *
-from .core import PriorDistribution
-from .problem import Problem
+from .core import PriorDistribution, Problem
 
 logger = logging.getLogger(__name__)
